# KAFKA-10470: Zstd upgrade and buffering