--- a/gradle/dependencies.gradle
+++ b/gradle/dependencies.gradle
@@ -116,7 +116,7 @@ versions += [
   testRetryPlugin: "1.1.6",
   zinc: "1.3.5",
   zookeeper: "3.5.8",
-  zstd: "1.4.5-8"
+  zstd: "1.4.5-10"
 ]
 libs += [
   activation: "javax.activation:activation:$versions.activation",