@yuzawa-san 2020-11-10
There is a newer zstd-jni version, so I'm going to push the newer one shortly after my local checks pass.